// src/app/models/reportCODConfirm.model.ts
export class ReportCODConfirm {
    id: number;
    shipmentNumber: string;
    requestShipmentId: number;
    cusNote: string;
    orderDate: Date;
    paymentTypeId: number;
    paymentTypeName: string;
    serviceId: number;
    serviceName: string;
    defaultPrice: number;
    fuelPrice: number;
    remoteAreasPrice: number;
    totalDVGT: number;
    vatPrice: number;
    totalPrice: number;
    shipmentStatusId: number;
    shipmentStatusName: string;
    deliverUserId: number;
    deliverUserName: string;
    senderName: string;
    senderPhone: string;
    companyFrom: string;
    pickingAddress: string;
    fromProvinceId: number;
    fromProvinceName: string;
    fromHubId: number;
    fromHubName: string;
    fromHubRoutingId: number;
    fromHubRoutingName: string;
    receiverName: string;
    receiverPhone: string;
    companyTo: string;
    shippingAddress: string;
    realRecipientName: string;
    endDeliveryTime: Date;
    toProvinceId: number;
    toProvinceName: string;
    toHubRoutingId: number;
    toHubRoutingName: string;
    totalCount: number;
    totalCOD: number;
    cod: number;
}